+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [2.4.75] - 2024-04-11
-### Added
+### Fixed
 - [Exchanges] Properly handle order not found errors
+- [Order] Rare synch issues on creation
+- [Order] Missing ungrouped stop orders when restarting
 
 ## [2.4.74] - 2024-04-07
 ### Added